修改变量类型;

+++ src/http/request.h
@@ -19,9 +19,9 @@
 #define GENE_REQUEST_PROPERTY_ATTR "_attr"
 
 extern zend_class_entry *gene_request_ce;
-zval * request_query(zend_ulong type, char * name, int len TSRMLS_DC);
+zval * request_query(zend_ulong type, char * name, size_t len TSRMLS_DC);
 void setVal(zend_ulong type, zval *value);
-zval *getVal(zend_ulong type, char *name, int len);
+zval *getVal(zend_ulong type, char *name, size_t len);
 
 #define GENE_REQUEST_IS_METHOD(ce, x) \
 PHP_METHOD(ce, is##x) {\
@@ -41,7 +41,7 @@ PHP_METHOD(ce, is##x) {\
 #define GENE_REQUEST_METHOD(ce, x, type) \
 PHP_METHOD(ce, x) { \
 	char *name = NULL; \
-	int name_len = 0;  \
+	size_t name_len = 0;  \
     zval *ret = NULL, *def = NULL; \
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "|sz", &name, &name_len, &def) == FAILURE) { \
 		return; \
